requirements.txt
crewai>=0.16.0
python-dotenv>=0.19.0
pymongo>=4.0.0
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.2
pydantic>=2.0
litellm>=1.0.0
pyyaml>=6.0
